64-bit patches
git-svn-id: https://g3d.svn.sourceforge.net/svnroot/g3d@1001 ef0428c2-0c16-0410-9f3f-d710ca1f8061

+++ G3D9/G3D.lib/include/G3D/FileSystem.h
@@ -517,7 +517,7 @@ public:
     static bool matches(const std::string& path, const std::string& pattern, bool caseSensitive = true);
 
     /** Replaces characters that are illegal in a filename with legal equivalents.*/
-    static std::string makeLegalFilename(const std::string& f, int maxLength = 100000);
+    static std::string makeLegalFilename(const std::string& f, size_t maxLength = 100000);
 };
 
 } // namespace G3D
